+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
-- Fix "start_date" and "end_date" timestamp [#52](https://github.com/rokwire/surveys-building-block/issues/52)
+- Fix "start_date" and "end_date" timestamp, "archived", "public" and set "complete" in the result [#52](https://github.com/rokwire/surveys-building-block/issues/52)
 
 ## [1.7.0] - 2024-007-24
 ### Added