+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.8.2] - 2024-007-26
+### Fixed
+- Fix Get /surveys [#58](https://github.com/rokwire/surveys-building-block/issues/58)
+
 ## [1.8.1] - 2024-007-26
 ### Fixed
 - Fix "public" and "archived" [#55](https://github.com/rokwire/surveys-building-block/issues/55)